other/models being tested/Python Web App for testing (flaskapp)/src/requirements.txt
Flask==2.1.1
gunicorn==20.1.0
firebase-admin==5.0.2
boto3==1.34.48


llama_index==0.10.19
numpy==1.24.4
python-dotenv==1.0.1
replicate==0.24.0
transformers==4.38.2
django-cors-headers
psycopg2-binary>=2.9
llama-index-embeddings-huggingface==0.1.4

firebase-admin==5.0.2